Fixed output header. Removed species field.
Change:
"entity1" to ":START(Entity-id)"
"entity2" to ":END(Entity-id)"
".ind" to "_ind"

This can be done directly on the output header string using the built in methods.
Ex. header.replace('entity1, ':START') and/or header.replace('.ind','_ind')

Add:
":float" to all predicate code columns.

The :START and :END tags are for the neo4j command line import tool. The The dot notation for the "ind" tags can lead to namespace issues for the cypher query language. The :float tag is needed for the import tool (import defaults to :string type).